+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `Client::createSubscription`
   - `Client::getSubscriptions`
 - `Client::runIterate` method (#45)
+- `Client::isConnected` method
 - `ReadValueId` wrapper class (#45)
 - Example `client_subscription.cpp` (#45)
 - `ByteString::fromBase64` and `ByteString::toBase64` methods